src/test/java/oop/recap2Shop/ProductRepoTest.java
package oop.recap2Shop;

import org.junit.jupiter.api.Test;

import java.util.ArrayList;
import java.util.Collections;
import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

class ProductRepoTest {

    @Test
    void getProducts() {
        //GIVEN
        ProductRepo repo = new ProductRepo();
        //WHEN
        List<Product> actual = repo.getProducts();
        //THEN
        List<Product> expected = new ArrayList<>();
        expected.add(new Product("1", "Apfel"));
        assertEquals(actual, expected);
    }

    @Test
    void getProductById() {
        //GIVEN
    ProductRepo repo = new ProductRepo();
        //WHEN
        Product actual = repo.getProductById("1");
        //THEN
        Product expected = new Product ("1", "Apfel");
        assertEquals(actual, expected);
    }

    @Test
    void addProduct() {
        //GIVEN
        ProductRepo repo = new ProductRepo();
        Product newProduct = new Product ("2", "Banane");
        //WHEN
        Product actual = repo.addProduct(newProduct);
        //THEN
        Product expected = new Product("2", "Banane");
        assertEquals(actual, expected);
        assertEquals(repo.getProductById("2"), expected);
    }

    @Test
    void removeProduct() {
        //GIVEN
        ProductRepo repo = new ProductRepo();
        //WHEN
        repo.removeProduct("1");
        //THEN
        assertNull(repo.getProductById("1"));
    }
}